types: extract store types to store.ts
